Allow Ctrl-b to work in nested tmux
When starting a nested tmux, do not revert the tmux prefix in the parent
tmux to the default (C-b) as it intercepts Ctrl-b from the nested tmux.
Instead, bind the tmux prefix to C-F12, which is unlikely to be used and
parallels the M-F12 binding used for active handler.

For example, to reproduce the problem:

1) Start tmux
2) Start vim in paraent tmux
   - Ctrl-f works to go forward a page
   - Ctrl-b works to go back a page
2) Create nested tmux (C-a C-s)
3) Start vim in nested tmux
   - Ctrl-f works to go forward a page
   - Ctrl-b does not work to go back a page